requirements.txt
Flask==2.1.0
gunicorn==19.9.0
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.54.0